requirements.txt
numpy
soundcard
pystray
Pillow
keyboard
PySide6
